docker/account_management/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary